Client for the Bittrex Crypto-Currency Exchange

This software is in no way affiliated, endorsed, or approved by the Bittrex crypto-currency exchange or any of its affiliates. It comes with absolutely no warranty and should not be used in actual trading unless the user can read and understand the source and knows what they are doing.

The bittrex package is an R implementation of the REST interface used by the Bittrex crypto-currency exchange (https://bittrex.com). It provides functions for all endpoints currently (as of May 16, 2017) supported by the exchange. This includes the ability to retrieve price, volume, and orderbook information as well as the ability to trade crypto-currencies.

Calls to the exchange are categorized as either public, which includes requests for price, volume, and order book information, and private, which includes all requests requiring an account including placing buy or sell orders. Public calls can be used immediately after installing the package. Private calls require creating an account at https://bittrex.com and creating API and secret keys with appropriate permissions.

Private calls retrieve the API and secret key using the BITTREX_API_KEY and BITTREX_SECRET_KEY environment variables. These may be set by the user before opening the R session or, they can be set using the bt_authenticate() function.

Public Function Calls

bt_api_check() check if the bittrex REST API is working

bt_getcurrencies() all supported currencies at Bittrex along with other meta data

bt_getmarkethistory() the latest trades that have occurred for a specified market

bt_getmarkets() the open and available trading markets at Bittrex along with other meta data

bt_getmarketsummary() the last 24 hours' summary of a specific market

bt_getmarketsummaries() the last 24 hours' summary of all active markets

bt_getorderbook() the orderbook for a given market

bt_getticker() the current tick values for a market

Private Function Calls

bt_authenticate() provide user authentication data

bt_buy() place a buy limit order

bt_cancel() cancel a buy or sell order

bt_getbalances() account balances for all currencies

bt_getbalance() account balance for a specified currency

bt_getdepositaddress() retrieve or generate an address for a specified currency

bt_getdeposithistory() retrieve your deposit history

bt_getopenorders() retrieve data for all open orders

bt_getorder() retrieve a single order by uuid

bt_getorderhistory() recent order history for an account

bt_getwithdrawalhistory() retrieve your withdrawal history

bt_sell() place a sell limit order

bt_withdraw() withdraw funds from your account

References

https://bittrex.com, https://github.com/kaneplusplus/bittrex